backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
simple-websocket==1.1.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.5
wsproto==1.3.2
